pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["restgdf*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
restgdf = ["py.typed"]

[project]
name = "restgdf"
version = "2.0.0"
description = "lightweight async Esri REST client with optional GeoPandas extras"
authors = [{ name = "Joshua Sundance Bailey" }]
readme = "README.md"
dependencies = [
    "aiohttp",
    "pydantic>=2.13.3,<3",
    "eval_type_backport; python_version < '3.10'",
    "requests",
]
license = "MIT"
license-files = ["LICENSE"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: AsyncIO",
    "Framework :: Pydantic",
    "Framework :: Pydantic :: 2",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
keywords = [
    "geopandas",
    "esri",
    "arcgis",
    "async",
    "aiohttp",
    "pydantic",
    "rest",
    "gis",
    "featureserver",
    "mapserver",
]
requires-python = ">=3.9"
[project.urls]
Homepage = "https://github.com/joshuasundance-swca/restgdf"
Documentation = "https://restgdf.readthedocs.io/"
Repository = "https://github.com/joshuasundance-swca/restgdf"
Issues = "https://github.com/joshuasundance-swca/restgdf/issues"
Changelog = "https://github.com/joshuasundance-swca/restgdf/blob/main/CHANGELOG.md"
"Security Policy" = "https://github.com/joshuasundance-swca/restgdf/security/policy"
"Ask DeepWiki" = "https://deepwiki.com/joshuasundance-swca/restgdf"
"llms.txt" = "https://restgdf.readthedocs.io/en/latest/llms.txt"


[project.optional-dependencies]
geo = [
    "geopandas",
    "pandas",
    "pyogrio",
]

resilience = [
    "stamina>=24.2",
    "aiolimiter>=1.1",
]

telemetry = [
    "opentelemetry-api>=1.20",
    "opentelemetry-instrumentation-aiohttp-client>=0.44b0",
]

dev = [
    "aioresponses>=0.7",
    "bumpver",
    "coverage",
    "hypothesis",
    "opentelemetry-instrumentation-aiohttp-client>=0.44b0",
    "opentelemetry-sdk>=1.20",
    "pre-commit",
    "pytest",
    "pytest-asyncio",
    "setuptools>=82.0.1",
]

doc = [
    "sphinx>=8.0,<10",
    "furo>=2024.8.6",
    "autodoc-pydantic>=2.2,<3",
    "myst-parser[linkify]>=5.0",
    "sphinx-copybutton>=0.5.2",
    "sphinx-design>=0.7",
    "sphinx-llm>=0.4",
]

[tool.bumpver]
current_version = "2.0.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "Bump version {old_version} -> {new_version}"
commit = true
tag_message = "{new_version}"
tag_scope = "default"
tag = true
push = true
# Stamps today's UTC date into CITATION.cff `date-released` before the
# bump commit. SemVer `version_pattern` has no date tokens, so this is
# the only way to keep `date-released` in sync via bumpver. The hook is
# exec'd directly by bumpver (see site-packages/bumpver/hooks.py); the
# file needs a shebang + executable bit on POSIX, and the standard .py
# association on Windows (py launcher).
pre_commit_hook = "scripts/bumpver_stamp_date.py"

[tool.bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']
"restgdf/__init__.py" = ['__version__ = "{version}"']
# Anchor the CITATION.cff pattern to avoid accidentally matching the
# `cff-version: 1.2.0` schema-version line, which contains `version: ...`
# as a substring. See https://github.com/mbarkhau/bumpver/issues/215.
"CITATION.cff" = ['^version: {version}$']


[tool.coverage.run]
branch = true
omit = ["*tests/*.py", "restgdf/app.py"]
command_line = "-m pytest"
source = ["restgdf"]

[tool.coverage.report]
show_missing = true
fail_under = 97
# Regexes for lines to exclude from consideration
exclude_also = [
    # Don't complain about missing debug-only code:
    "def __repr__",
    "if self\\.debug",

    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",

    # Don't complain about abstract methods, they aren't run:
    "@(abc\\.)?abstractmethod",

    # Don't complain about type-checking-only imports / overloads:
    "if TYPE_CHECKING:",
    "@overload",

    # Don't complain about Protocol / ABC method stubs:
    "^\\s*\\.\\.\\.\\s*$",
    ]

ignore_errors = true

[tool.coverage.html]
directory = "coverage_html_report"
skip_covered = false
skip_empty = true
title = "restgdf coverage report"

[tool.pytest.ini_options]
pythonpath = "."
testpaths = "tests"
addopts = "--strict-markers"
asyncio_mode = "strict"
markers = [
    "network: tests that hit live external ArcGIS services",
    "characterization: tests that pin current behavior to guard refactors",
    "compat: backward-compatibility tests for preserved import/patch paths",
]

[tool.mypy]

[[tool.mypy.overrides]]
module = ["restgdf._client.*", "restgdf._models.*", "restgdf.compat"]
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
warn_return_any = true
warn_unused_ignores = true
no_implicit_optional = true
strict_equality = true

[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
ignore = ["E501"]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B311"]
